+++ src/const.js
@@ -1,6 +1,5 @@
 export const TRIP_POINTS_MAP = new Map([[`taxi`, `to`], [`bus`, `to`], [`train`, `to`], [`ship`, `to`], [`transport`, `to`],
   [`drive`, `to`], [`flight`, `to`], [`check-in`, `in`], [`sightseeing`, `in`], [`restaurant`, `in`]]);
-export const ADDITIONAL_OPTIONS = [`Order Uber`, `Add meal`, `Add luggage`, `Travel by train`];
 export const TRIP_POINT_TRANSFER_TYPES = [`taxi`, `bus`, `train`, `ship`, `transport`, `drive`, `flight`];
 export const TRIP_POINT_ACTIVITIES_TYPE = [`check-in`, `sightseeing`, `restaurant`];
