+++ src/presenter/filter.js
@@ -37,18 +37,6 @@ export default class Filter {
     remove(prevFilterComponent);
   }
 
-  _handleModelEvent() {
-    this.init();
-  }
-
-  _handleFilterTypeChange(filterType) {
-    if (this._currentFilter === filterType) {
-      return;
-    }
-
-    this._filterModel.setFilter(UpdateType.MINOR, filterType);
-  }
-
   _getFilters() {
     return [
       {
@@ -64,5 +52,17 @@ export default class Filter {
         name: `FUTURE`,
       },
     ];
+  }
+
+  _handleModelEvent() {
+    this.init();
+  }
+
+  _handleFilterTypeChange(filterType) {
+    if (this._currentFilter === filterType) {
+      return;
+    }
+
+    this._filterModel.setFilter(UpdateType.MINOR, filterType);
   }
 }
