+++ src/utils/common.js
@@ -10,17 +10,3 @@ export const generateRandomListItem = (list) => {
 
   return list[randomIndex];
 };
-
-export const updateItem = (items, update) => {
-  const index = items.findIndex((item) => item.id === update.id);
-
-  if (index === -1) {
-    return items;
-  }
-
-  return [
-    ...items.slice(0, index),
-    update,
-    ...items.slice(index + 1)
-  ];
-};
