+++ src/view/abstract.js
@@ -10,10 +10,6 @@ export default class Abstract {
     this._callback = {};
   }
 
-  _getTemplate() {
-    throw new Error(`Abstract method not implemented: getTemplate`);
-  }
-
   getElement() {
     if (!this._element) {
       this._element = createElement(this._getTemplate());
@@ -24,5 +20,9 @@ export default class Abstract {
 
   removeElement() {
     this._element = null;
+  }
+
+  _getTemplate() {
+    throw new Error(`Abstract method not implemented: getTemplate`);
   }
 }
