+++ src/view/point-list-element.js
@@ -1,11 +1,11 @@
 import AbstractView from "./abstract";
 
 export default class PointListElement extends AbstractView {
-  _createPointListElement() {
-    return (`<li class="trip-events__item"></li>`);
-  }
-
   _getTemplate() {
     return this._createPointListElement();
+  }
+
+  _createPointListElement() {
+    return (`<li class="trip-events__item"></li>`);
   }
 }
