+++ src/view/statistics.js
@@ -1,38 +1,8 @@
 import Chart from "chart.js";
 import ChartDataLabels from 'chartjs-plugin-datalabels';
 import SmartView from "./smart.js";
+import {formatDatesDifference} from "../utils/point";
 import {TRIP_POINT_TYPES} from "../mock/trip-point";
-
-const formatDatesDiff = (dateDiff) => {
-  const daysDiff = Math.floor(dateDiff / (24 * 60 * 60 * 1000));
-  dateDiff -= daysDiff * (24 * 60 * 60 * 1000);
-  const hoursDiff = Math.floor(dateDiff / (1000 * 60 * 60));
-  dateDiff -= hoursDiff * (60 * 60 * 1000);
-  const minutesDiff = Math.floor(dateDiff / (60 * 1000));
-
-  let dateDiffString = ``;
-  if (daysDiff) {
-    if (daysDiff < 10) {
-      dateDiffString += `0` + daysDiff + `D `;
-    } else {
-      dateDiffString += daysDiff + `D `;
-    }
-  }
-  if (hoursDiff) {
-    if (hoursDiff < 10) {
-      dateDiffString += `0` + hoursDiff + `H `;
-    } else {
-      dateDiffString += hoursDiff + `H `;
-    }
-  }
-  if (minutesDiff < 10) {
-    dateDiffString += `0` + minutesDiff + `M`;
-  } else {
-    dateDiffString += minutesDiff + `M`;
-  }
-
-  return dateDiffString;
-};
 
 const renderMoneyChart = (moneyCtx, data) => {
   let map = new Map([
@@ -256,7 +226,7 @@ const renderTimeChart = (timeCtx, data) => {
           color: `#000000`,
           anchor: `end`,
           align: `start`,
-          formatter: (val) => `${formatDatesDiff(val)}`
+          formatter: (val) => `${formatDatesDifference(val)}`
         }
       },
       title: {
