+++ src/view/trip-events-list.js
@@ -1,13 +1,13 @@
 import AbstractView from "./abstract";
 
 export default class TripEventsList extends AbstractView {
+  _getTemplate() {
+    return this._createTripEventsList();
+  }
+
   _createTripEventsList() {
     return (
       `<ul class="trip-events__list"></ul>`
     );
-  }
-
-  _getTemplate() {
-    return this._createTripEventsList();
   }
 }
